requirements.txt
# core
# jentic-sdk>=0.1.0  # TODO: add when available
openai>=1.0
pydantic>=2.0
litellm

# Discord bot dependencies
PyNaCl>=1.5.0        # for Discord signature verification
requests>=2.28.0     # for Discord API calls
